+++ browsertrace/__init__.py
@@ -2,5 +2,5 @@
 
 from .tracer import Run, Tracer, trace
 
-__version__ = "0.1.4"
+__version__ = "0.1.5"
 __all__ = ["Tracer", "Run", "trace"]